Moved post report to moderation module
