--- src/posts/dtos/reportPostPayload.dto.ts
@@ -1,13 +1,0 @@
-import { ApiProperty } from "@nestjs/swagger";
-
-export class ReportPostPayloadDto {
-    @ApiProperty({ type: String, format: "uuid" })
-    postId: UUID;
-
-    @ApiProperty({ type: String, minLength: 5, maxLength: 500 })
-    reason: string;
-
-    constructor(partial?: Partial<ReportPostPayloadDto>) {
-        Object.assign(this, partial);
-    }
-}
