--- src/posts/services/postReport/postsReport.service.interface.ts
@@ -1,8 +1,0 @@
-import { ReportPostPayloadDto } from "../../dtos";
-import { ReportedProps } from "../../../users/models/toPost";
-
-export interface IPostsReportService {
-    reportPost(reportPostPayload: ReportPostPayloadDto): Promise<void>;
-
-    getReportsForPost(postId: string): Promise<ReportedProps[]>;
-}
